requirements.txt
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
Django==5.2.6
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
narwhals==2.4.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.0
pyarrow==21.0.0
pycryptodomex==3.22.0
pydeck==0.9.1
pygame==2.6.1
PyOpenGL==3.1.9
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.49.1
psycopg
psycopg2
gunicorn
dj-database-url
pyserial
